Move CheckLeafPartitionCount into bulk-import-common
