+++ java/bulk-import/bulk-import-runner/src/main/java/sleeper/bulkimport/job/runner/BulkImportJobDriver.java
@@ -36,7 +36,6 @@ import sleeper.core.record.process.RecordsProcessedSummary;
 import sleeper.ingest.job.status.IngestJobStatusStore;
 import sleeper.ingest.status.store.job.IngestJobStatusStoreFactory;
 import sleeper.statestore.StateStore;
-import sleeper.statestore.StateStoreException;
 import sleeper.statestore.StateStoreProvider;
 
 import java.io.IOException;
@@ -47,7 +46,7 @@ import java.time.Duration;
 import java.time.Instant;
 import java.util.function.Supplier;
 
-import static sleeper.bulkimport.job.runner.CheckLeafPartitionCount.hasMinimumPartitions;
+import static sleeper.bulkimport.CheckLeafPartitionCount.hasMinimumPartitions;
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.BULK_IMPORT_BUCKET;
 
 /**
@@ -97,14 +96,13 @@ public class BulkImportJobDriver {
                 getTime);
     }
 
-    public void run(BulkImportJob job, String taskId) throws IOException, StateStoreException {
+    public void run(BulkImportJob job, String taskId) throws IOException {
         Instant startTime = getTime.get();
         LOGGER.info("Received bulk import job with id {} at time {}", job.getId(), startTime);
         LOGGER.info("Job is {}", job);
 
 
-        if (!hasMinimumPartitions(stateStoreProvider, tablePropertiesProvider, job.getTableName())) {
-            LOGGER.info("Skipping job {}", job.getId());
+        if (!hasMinimumPartitions(stateStoreProvider, tablePropertiesProvider, job)) {
             return;
         }
         statusStore.jobStarted(taskId, job.toIngestJob(), startTime);
